feat(consensus): active dynamic validators at allegretto-time=0 (#1249)
This patch ensures that dynamic validator logic is active when
`allegretto-time = 0` is set in genesis.

The xtask tool is changed to always write the initial DKG outcome into
the genesis header and the initial validator set into the genesis
validator config contract. To simulate a hardfork on testnet, the escape
hatches `--no-validators-in-genesis` and `--no-dkg-in-genesis` are
provided.

The test setup was expanded to also cover the cases where
`allegretto-time = 0` to test the specific code paths of the actor.

Closes #1087
Closes #1211

---------

Co-authored-by: joshieDo <93316087+joshieDo@users.noreply.github.com>